app/requirements.txt
flask
boto3
flask-jwt-extended
flask-limiter
python-dotenv
redis
cryptography
phonenumbers
flask-wtf
werkzeug
gunicorn
